Rework page routing.
The model now keeps a currentPage property, where Page is an ADT, and
page routing runs off of that.  There are still hooks into the browser
history mechanism, but we go through a complex path to turn a URL-change
event into an arrived-on-page event.  The idea is that we can further hook
into this to do things like fetch a stop list on arriving on the route
detail page.