Reuse StopTime identifiers to cut down memory use.
Every StopTime has a reference to a Stop and a Time, and there are many many
of these.  We can save maybe a quarter of memory use on the MBTA feed by
looking up identifiers on the referenced objects and so have one Identifier
per Stop, and not one (Identifier Stop) per StopTime (many duplicated).