Move all of the updater logic into TellMeT.UI.Update.
The real rationale for this is to decouple the fetching cycle from the
view code a little more: the route page is getting complex enough that
the split Redux encourages where you have separate action constructors,
update functions, selectors on the model, and views actually starts to
make sense.

This also allows us to enforce that every action does get processed
in the update function, and to get rid of the "if..." variants in the
action typeclasses.  The unified update file is actually a reasonable size.